cloud_run/rembg_service/requirements.txt
flask==3.0.3
gunicorn==22.0.0
firebase-admin==6.5.0
google-cloud-storage==2.18.2
rembg==2.0.61
onnxruntime==1.19.2
